+++ src/lib.rs
@@ -5,14 +5,11 @@
 //!
 //! # Cargo features
 //!
-//! Default features: log, upload, rustls, human-panic
+//! Default features: log, upload, human-panic
 //!
 //! - log: Configures pretty-env-logger, even though maturin doesn't use logging itself.
 //!
-//! - upload: Uses reqwest to add the upload command.
-//!
-//! - rustls: Makes reqwest use the rustls stack so that we can build maturin in a CentOS 6
-//! docker container and which maturin itself manylinux compliant.
+//! - upload: Uses ureq to add the upload command.
 //!
 //! - human-panic: Adds https://github.com/rust-clique/human-panic
 //!